Adapt and tweak workflows and tools to work with current cwltool
Some necessary changes for the correct execution of the pipelines:
- valueFrom: $(NUMBER) --> ${return NUMBER}. Not sure why this is necessary
- secondaryFiles are expected to be in the same location than the primary files. In
previous cwltool versions, all subworkflow files will be copied over a shared output
directory, grouping all output files from one subworkflow together. That is not the
case anymore. As a consequence, index files and other secondaryFiles need to be
"attached" to the principal files, to ensure their co-localization. That affects programs
like samtools-index, which now returns the input BAM file with the index file as a
secondaryFile.